[Merge] integrate changes from 2.4 branch (r4418)

+++ branches/avidemux_2.5_branch_gruntster/avidemux/ADM_UIs/ADM_QT4/src/T_flyDialogQt4.cpp
@@ -42,7 +42,7 @@ bool FlyDialogEventFilter::eventFilter(QObject *obj, QEvent *event)
 		UI_getPhysicalScreenSize(parent, &screenWidth, &screenHeight);
 		flyDialog->recomputeSize();
 		QCoreApplication::processEvents();
-		parent->move((screenWidth - parent->frameSize().width()) / 2, (screenHeight - parent->frameSize().height()) / 2);
+		parent->move((((int)screenWidth) - parent->frameSize().width()) / 2, (((int)screenHeight) - parent->frameSize().height()) / 2);
 	}
 }
